requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
httpx==0.28.0
beautifulsoup4==4.12.3
google-generativeai==0.8.3
python-dotenv==1.0.1
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.12
